factor out `onDrop`
