+++ editor/src/components/canvas/event-helpers.test-utils.tsx
@@ -522,7 +522,7 @@ export function pressKey(
 
 // https://github.com/testing-library/react-testing-library/issues/339
 export function makeDragEvent(
-  type: 'drag' | 'drop',
+  type: 'dragover' | 'drop',
   target: Element | Node,
   clientCoords: { x: number; y: number },
   fileList: Array<File>,
@@ -535,7 +535,7 @@ export function makeDragEvent(
     cancelable: true,
   }
   const fileDropEvent =
-    type === 'drop' ? createEvent.drop(target, opts) : createEvent.drag(target, opts)
+    type === 'drop' ? createEvent.drop(target, opts) : createEvent.dragOver(target, opts)
 
   Object.defineProperty(fileDropEvent, 'dataTransfer', {
     value: {
